Doc-Processing-Pipeline-for-knowledge-base/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
psycopg2-binary
SQLAlchemy
pydantic
python-multipart
Jinja2 